add option -dummy to run_tests
the script "run_tests" now allows for an extra option "-dummy". If this
is given, a test-directory with prefix "dummy_" is created. The
intention is that these dummy-directories can easily be deleted
afterwards and are used for local, very temporary testing. Real tests
should be run without this option. Then the old naming sheme applies and
the results are automatically staged for git (but not commited
automatically).